test(gateway): split provider-provisioning replay coverage
Direct commit 4d648892909 pushed provider-provisioning.test.ts to 1091
lines, past the 1000-line max-lines cap, breaking check-lint-core-2 on
main. Move the replay/restart-adoption/serialization tests into
provider-provisioning.replay.test.ts (same suite setup via
service.test-support), matching the repo's split-not-suppress policy.
